requirements.txt
Flask
gunicorn
gspread
oauth2client
apscheduler
requests
yagmail
groq
python-dotenv